+++ awb_amount_based_approvals/models/approval_category.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from odoo import api, fields, models, _
+
+class ApprovalType(models.Model):
+    _inherit = 'approval.category'
+
+    rule_ids = fields.Many2many('approval.rule', String='Rules')