+++ awb_generic_sms/docs/CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.2.2] - 2021-07-02
+   - Scheduled Action Payment Reminder due date criteria
+   - Scheduled Action Disconnection Notice due date criteria
+   - Include Payment Notification Template update
+
+## [13.0.1.2.1] - 2021-06-23
+   - New validations
+   - New sms state (sent, failed, failed - missing parameters)
+   - Added sms sending source
+   - New features
+
+## [13.0.1.1.1] - 2021-06-10
+   - Removed atm_ref prefix slicer
+
 ## [13.0.1.1.0-beta] - 2021-05-19
    - Payment Notification Template(Account No -> Account Name)
    - Scheduled Actions Fixes